Made 1 minor change to test git commit messages
I made this change simply to test the git commit messages and see what it looks like to have it on 2 lines so that in the future I can do it correctly and not incorrectly. That is why I am typing so much so I can see how a very long commit message will look. Not that this is super long, but long enough to test out what will happend if a somewhat long message is used in the git commit. I suppose this is long enough now and I will leave it here.